+++ CHANGELOG.md
@@ -6,7 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [0.6.0] - 2026-04-02
+## [0.6.1] - 2026-04-13
+
+### Changed
+- The `nxlib.nxopen.part.part_context` function signature was updated to allow more explicit control over what happens when exiting the context manager. The `close_when_true` argument still
+works but raises a deprecation warning.
+
+### Fixed
+- Parts are properly closed when the `nxlib.nxopen.part.part_context` complete. Error messages such as `In non-interactive mode.  Export directory C:\Users\2343177_1 is deleted.` at the end of journal execution should not longer be present.
+- Opening an assembly with `open_assembly=False` (default) explicitly sets the load options to only load the base part; the entire assembly was being opened previously.
+
+## [0.6.0] - 2026-03-27
 
 ### Added
 - Added the `nxlib.geometry.CartesianCoordinateSystem` class, which mirrors the `NXOpen.CartesianCoordinateSystem` class.